Move serialize from cannon to op-service
