+++ op-service/serialize/binary.go
