+++ op-service/serialize/binary_test.go
