+++ op-service/serialize/detect_test.go
