+++ op-service/serialize/reader_test.go
