+++ op-service/serialize/writer.go
