+++ op-service/serialize/writer_test.go
